Cleanup unused, refactor graphData method
