+++ src/app/calculator/furnaces/o2-enrichment/o2-enrichment-copy-table/o2-enrichment-copy-table.component.html
@@ -1,6 +1,5 @@
 <div *ngIf="outputs.length > 0">
   <div class="hidden-copy-table">
-    <!-- <table [ngStyle]="{'visibility': clickedBtn == true? 'visible' : 'collapse'}" #enrichmentTable> -->
       <table #enrichmentTable>
       <tbody>
         <tr>
