--- src/app/calculator/furnaces/o2-enrichment/o2-enrichment-graph/o2-enrichment-graph.component.css
@@ -1,5 +1,0 @@
-
-
-.expanded-chart {
-  padding: 5px;
-}
