--- src/app/calculator/furnaces/o2-enrichment/o2-enrichment-graph/o2-enrichment-graph.component.html
@@ -1,49 +1,0 @@
-<!-- need a new id container to handle resizing events properly
-      conditional 'expanded-chart' class will handle creating the floating container -->
-<div #ngChartContainer class="panel-chart-container" [ngClass]="{'expanded-chart': expanded}">
-
-  <div *ngIf="expanded" class="row">
-    <div class="col-12">
-      <div class="graph-btn-container">
-        <span class="fa fa-compress" *ngIf="expanded" (click)="contractChart()" (mouseenter)="initTooltip('btnCollapseChart')"
-          (mouseleave)="hideTooltip('btnCollapseChart')"></span>
-        <app-simple-tooltip *ngIf="expanded" [message]="'Collapse graph'" [hover]="hoverBtnCollapse" [display]="displayCollapseTooltip"
-          [adjustLeft]="20" [adjustPercent]="false"></app-simple-tooltip>
-      </div>
-    </div>
-  </div>
-
-  <div class="d-flex w-100 justify-content-center align-items-center" [ngClass]="{'h-100': !expanded, 'h-90': expanded}">
-    <div class="row m-0">
-      <div class="col-12 p-0">
-        <div #ngChart></div>
-      </div>
-    </div>
-  </div>
-
-  <div class="col-md-12">
-    <div class="graph-btn-container">
-      <span class="fa fa-download" (click)="downloadChart()" (mouseenter)="initTooltip('btnExportChart')" (mouseleave)="hideTooltip('btnExportChart')"></span>
-      <app-simple-tooltip [message]="'Download graph as image'" [hover]="hoverBtnExport" [display]="displayExportTooltip"
-        [adjustLeft]="20" [adjustPercent]="false"></app-simple-tooltip>
-    </div>
-    <div class="graph-btn-container">
-      <span class="fa fa-bar-chart" (click)="toggleGrid()" (mouseenter)="initTooltip('btnGridLines')" (mouseleave)="hideTooltip('btnGridLines')"></span>
-      <app-simple-tooltip [message]="'Toggle graph grid lines'" [hover]="hoverBtnGridLines" [display]="displayGridLinesTooltip"
-        [adjustLeft]="59" [adjustPercent]="false"></app-simple-tooltip>
-    </div>
-    <div class="graph-btn-container">
-      <span class="fa fa-expand" *ngIf="!expanded" (click)="expandChart()" (mouseenter)="initTooltip('btnExpandChart')"
-        (mouseleave)="hideTooltip('btnExpandChart')"></span>
-      <app-simple-tooltip *ngIf="!expanded" [message]="'Expand graph'" [hover]="hoverBtnExpand" [display]="displayExpandTooltip"
-        [adjustLeft]="101" [adjustPercent]="false"></app-simple-tooltip>
-    </div>
-  </div>
-
-</div>
-
-<div class="d-flex w-100 justify-content-center pl-2 pr-2" *ngIf="!expanded && rowData.length > 0">
-  <app-exportable-table class="exportable-table-container" [settings]="settings" [columnTitles]="columnTitles"
-    [rowData]="rowData" [keyColors]="keyColors" (highlightPoint)="highlightPoint($event)" (unhighlightPoint)="unhighlightPoint($event)"
-    (deleteFromTable)="deleteFromTable($event)"></app-exportable-table>
-</div>
