+++ src/app/calculator/furnaces/o2-enrichment/o2-enrichment-help/o2-enrichment-help.component.ts
@@ -8,7 +8,6 @@ import { Subscription } from 'rxjs';
   styleUrls: ['./o2-enrichment-help.component.css']
 })
 export class O2EnrichmentHelpComponent implements OnInit {
-  @Input()
   currentField: string;
   currentFieldSub: Subscription;
