+++ src/app/calculator/furnaces/o2-enrichment/o2-enrichment.component.html
@@ -37,7 +37,7 @@
       </div>
     </div>
     <div class="d-flex mr-2 ml-2">
-      <app-o2-enrichment-help *ngIf="tabSelect == 'help'" [currentField]="currentField"></app-o2-enrichment-help>
+      <app-o2-enrichment-help *ngIf="tabSelect == 'help'"></app-o2-enrichment-help>
       <div class="d-flex flex-column w-100 align-items-center p-2 pt-4" *ngIf="tabSelect == 'results'"
         id="o2EnrichmentGraph">
         <app-enrichment-graph *ngIf="tabSelect == 'results'" class="w-100" [settings]="settings">
