+++ src/app/calculator/furnaces/o2-enrichment/o2-enrichment.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { O2EnrichmentHelpComponent } from '../o2-enrichment/o2-enrichment-help/o2-enrichment-help.component';
 import { O2EnrichmentComponent } from '../o2-enrichment/o2-enrichment.component';
-import { O2EnrichmentGraphComponent } from '../o2-enrichment/o2-enrichment-graph/o2-enrichment-graph.component';
 import { O2EnrichmentFormComponent } from '../o2-enrichment/o2-enrichment-form/o2-enrichment-form.component';
 import { O2EnrichmentService } from './o2-enrichment.service';
 import { ExportableResultsTableModule } from '../../../shared/exportable-results-table/exportable-results-table.module';
@@ -26,7 +25,6 @@ import { O2EnrichmentCopyTableComponent } from './o2-enrichment-copy-table/o2-en
   declarations: [
     O2EnrichmentFormComponent,
     O2EnrichmentComponent,
-    O2EnrichmentGraphComponent,
     O2EnrichmentHelpComponent,
     EnrichmentGraphComponent,
     O2EnrichmentResultsComponent,
